Relocate select table store-related files into global stores directory
