+++ src/components/tables/AccessGrants/index.tsx
@@ -7,8 +7,8 @@ import Rows, {
 import EntityTable from 'components/tables/EntityTable';
 import { useMemo } from 'react';
 import { SelectTableStoreNames } from 'stores/names';
+import useTableState, { TablePrefix } from 'stores/Tables/hooks';
 import TableHydrator from 'stores/Tables/Hydrator';
-import useTableState, { TablePrefix } from '../hooks';
 
 interface Props {
     tablePrefix: TablePrefix;
