+++ src/components/tables/Connectors/index.tsx
@@ -5,8 +5,8 @@ import EntityTable from 'components/tables/EntityTable';
 import { useMemo } from 'react';
 import { CONNECTOR_NAME } from 'services/supabase';
 import { SelectTableStoreNames } from 'stores/names';
+import useTableState from 'stores/Tables/hooks';
 import TableHydrator from 'stores/Tables/Hydrator';
-import useTableState from '../hooks';
 
 function ConnectorsTable() {
     const {
