+++ src/components/tables/Materializations/index.tsx
@@ -4,9 +4,9 @@ import EntityTable from 'components/tables/EntityTable';
 import Rows from 'components/tables/Materializations/Rows';
 import { useMemo } from 'react';
 import { SelectTableStoreNames } from 'stores/names';
+import useTableState from 'stores/Tables/hooks';
 import TableHydrator from 'stores/Tables/Hydrator';
-import StatsHydrator from '../Captures/StatsHydrator';
-import useTableState from '../hooks';
+import StatsHydrator from 'stores/Tables/StatsHydrator';
 import useMaterializationColumns from './useMaterializationColumns';
 
 function MaterializationsTable() {
