+++ src/components/tables/RowActions/Materialize.tsx
@@ -1,14 +1,14 @@
 import { Button } from '@mui/material';
 import { authenticatedRoutes } from 'app/routes';
-import {
-    SelectableTableStore,
-    selectableTableStoreSelectors,
-} from 'components/tables/Store';
 import { useZustandStore } from 'context/Zustand/provider';
 import { GlobalSearchParams } from 'hooks/searchParams/useGlobalSearchParams';
 import { FormattedMessage } from 'react-intl';
 import { useNavigate } from 'react-router';
 import { SelectTableStoreNames } from 'stores/names';
+import {
+    SelectableTableStore,
+    selectableTableStoreSelectors,
+} from 'stores/Tables/Store';
 import { getPathWithParams } from 'utils/misc-utils';
 
 interface Props {
