+++ src/components/tables/cells/stats/Header.tsx
@@ -1,12 +1,12 @@
 import { TableCell } from '@mui/material';
 import DateFilter from 'components/tables/Filters/Date';
-import {
-    SelectableTableStore,
-    selectableTableStoreSelectors,
-} from 'components/tables/Store';
 import { useTenantDetails } from 'context/fetcher/Tenant';
 import { useZustandStore } from 'context/Zustand/provider';
 import { SelectTableStoreNames } from 'stores/names';
+import {
+    SelectableTableStore,
+    selectableTableStoreSelectors,
+} from 'stores/Tables/Store';
 import { hasLength } from 'utils/misc-utils';
 
 interface Props {
