+++ src/stores/Tables/StatsHydrator.tsx
@@ -1,11 +1,11 @@
-import {
-    SelectableTableStore,
-    selectableTableStoreSelectors,
-} from 'components/tables/Store';
 import { useTenantDetails } from 'context/fetcher/Tenant';
 import { useZustandStore } from 'context/Zustand/provider';
 import { useEffect } from 'react';
 import { SelectTableStoreNames } from 'stores/names';
+import {
+    SelectableTableStore,
+    selectableTableStoreSelectors,
+} from 'stores/Tables/Store';
 import { BaseComponentProps } from 'types';
 import { hasLength } from 'utils/misc-utils';
